Add files via upload
The structure of the connections will now be circular. If the Master dies, a Client will become the new Master.
The Master and the Client share the same file: component.py
TO DO: Add the WatchDog thread for detecting a dead component. Add the process pair solution. Add the Main thread.